--- src/main/java/com/example/team_12_be/post/domain/ProjectQueryRepository.java
@@ -1,4 +1,0 @@
-package com.example.team_12_be.post.domain;
-
-public interface ProjectQueryRepository {
-}
