--- src/main/java/com/example/team_12_be/post/domain/comment/ProjectCommentRepository.java
@@ -1,9 +1,0 @@
-package com.example.team_12_be.post.domain.comment;
-
-import java.util.Optional;
-
-public interface ProjectCommentRepository {
-    ProjectComment save(ProjectComment projectComment);
-
-    Optional<ProjectComment> findById(Long id);
-}
