--- src/main/java/com/example/team_12_be/post/repository/ProejctJpaRepository.java
@@ -1,8 +1,0 @@
-package com.example.team_12_be.post.repository;
-
-import com.example.team_12_be.post.domain.Project;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProejctJpaRepository extends JpaRepository<Project, Long> {
-
-}
