+++ src/main/java/com/example/team_12_be/project/domain/ProjectLink.java
@@ -1,0 +1,28 @@
+package com.example.team_12_be.project.domain;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor
+@Getter
+public class ProjectLink {
+
+    @Id
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "project_id")
+    private Project project;
+
+    // TODO : 이 부분, 어떻게 처리할 지 논의. 데이터셋으로 넣어둔다?
+    private String siteType; // figma ...
+
+    private String url;
+
+}
