+++ src/main/java/com/example/team_12_be/project/domain/ProjectQueryRepository.java
@@ -1,0 +1,4 @@
+package com.example.team_12_be.project.domain;
+
+public interface ProjectQueryRepository {
+}
