+++ src/main/java/com/example/team_12_be/project/domain/ProjectRepository.java
@@ -1,0 +1,15 @@
+package com.example.team_12_be.project.domain;
+
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface ProjectRepository {
+
+    Project save(Project project);
+
+    void deleteById(Long projectId);
+
+    Optional<Project> findById(Long projectId);
+}
