+++ src/main/java/com/example/team_12_be/project/domain/comment/ProjectCommentRepository.java
@@ -1,0 +1,9 @@
+package com.example.team_12_be.project.domain.comment;
+
+import java.util.Optional;
+
+public interface ProjectCommentRepository {
+    ProjectComment save(ProjectComment projectComment);
+
+    Optional<ProjectComment> findById(Long id);
+}
