+++ src/main/java/com/example/team_12_be/project/repository/ProejctJpaRepository.java
@@ -1,0 +1,8 @@
+package com.example.team_12_be.project.repository;
+
+import com.example.team_12_be.project.domain.Project;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProejctJpaRepository extends JpaRepository<Project, Long> {
+
+}
